pyproject.toml
[project]
name = "gamer-news"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.10.5",
    "beautifulsoup4>=4.12.3",
    "feedparser>=6.0.11",
    "notebook>=7.2.2",
    "openai>=1.45.0",
    "pytest>=8.3.3",
    "python-fasthtml>=0.6.0",
]
